requirements.txt
langchain-core==0.3.6
langchain-openai==0.2.1
langchain-experimental==0.3.2
langchain-community==0.3.1
langchain==0.3.1
python-dotenv==1.0.1
langgraph==0.2.28
openai>=1.0.0
